NewAppFromConfig(): more flexible way to configure an `App`
The old code always created URIs with the following patterns:

* `http[s]://{ip}:{port}/api/[…]`
* `ws[s]://{ip}:{port}/api/websocket`

where `{ip}` and `{port}` are externally-specified parameters. But if,
for example, the code is running within a supervised container, then
it is supposed to connect via the proxy at

* `http://supervisor/core/api/[…]`
* `ws://supervisor/core/websocket` or `ws://supervisor/core/api/websocket`

It is not possible to generate such URIs with the existing
code (because of the extra `core/` in the path).

So instead, continue a little bit further with the idea of
https://github.com/saml-dev/gome-assistant/pull/17: add a second way
to initialize the app, where the caller can specify the URIs itself.
This is implemented via a new constructor function
`NewAppFromConfig()`, which takes a new configuration type,
`NewAppConfig`, as argument. Rewrite `NewApp()` to call
`NewAppFromConfig()`.