websocket.NewConn(): new function
The old code required the caller to create its own
`gorilla/websocket.Conn` using `websocket.ConnectionFromUri()`, and
then wrap that object in a `WebsocketConn` instance itself. But the
caller doesn't even need to know about the `gorilla/websocket.Conn`
anymore. So provide a new function, `websocket.NewConn()`, which opens
the connection and returns a `websocket.WebsocketConn` that is ready
to use.