Change the service calling convention
When calling a service, the caller isn't interested in the whole
result message. All they really need to know is if there was an error,
and if not, what was the result (without the message envelope).

Among other things, the makes it easier to handle RPC errors more
systematically.

So, change the low-level `App.Call*()` methods to take an additional
pointer to an object as argument. If the call is successful, the
`result` member of the result message will be deserialized into this
object using `json.Unmarshal()`.

If, on the other hand, the result has `success == false`, then convert
this into a `ResultError` and return that.

Change the concrete service types to return only the `result` part of
the result message. In most cases, this is declared as type `any`. If
people go to the trouble of figuring out the structure of the methods'
responses, these return types can be made more specific.