ApplyChange(): function to apply changes to an arbitrary entity
Change `CompressedStateChangedMessage.Apply()` from a method into a
generic top-level function, which can take an arbitrary `Entity` type
as argument and apply the change to it. This involves some type
conversion, which is done by serializing to and from JSON a couple of
times, but at least the application code doesn't need to be written
separately for each type.