Make it easier to compile the example app
When getting started with a new project, it is really convenient to
have a simple example that actually compiles. But the example app
currently doesn't compile because it is missing definitions for
entities that it depends on. Moreover, normal users can't generate
those entity definitions using `cmd/generate`, because their setup is
unlikely to have entities with the required names.

So:

* Include a hand-coded `example/entities/entities.go` file with the
  contents needed to compile the example app. To help prevent users
  from wondering why this doesn't work in their setup, include an
  `init()` method in this file that logs a warning explaining that
  example entities are in use, and pointing the user to `README.md`
  for instructions describing how to generate entities for their own
  setup.

* In `example/example.go`, uncomment the import of `example/entities`,
  so that the example app can actually be compiled.

* Change `example/gen.yaml`, to make it look more plausibly like the
  config that would be needed to generate the example app's own
  entities. Add a comment to this file, too, with a pointer to
  `README.md`.