websocket.ConnectionFromUri(): take a `ctx` parameter
Change `websocket.ConnectionFromUri()` to take a context parameter
rather than creating one internally and returning it to the caller.
This puts the caller in charge of the context, which makes more sense.

Note that this creates some warnings at the caller, because the
context is sometimes leaked (which was already the case, but not it's
obvious enough to be detected via static analysis). That will be fixed
in a moment.