App.ctx: store a long-lived context here
Long term it might or might not make sense for `App` to have `ctx` and
`ctxCancel` members. For example, right now its `ctx` member is
unused. But it probably makes sense for them to exist in some form.

But what _certainly_ doesn't make sense is storing a context there
that has a 3 second timeout, which is what has been done until now.
Three seconds is a reasonable timeout for connecting to the websocket,
but not for the lifetime of the `App`. So create a separate
`connContext` with a 3s timeout to use for connecting, and cancel that
context unconditionally. (This also removes the warnings about
possibly leaking the context on error paths.)

Store the caller's original `ctx` (made cancelable) to `App.ctx`. It's
still not used, but at least it won't cancel itself after only 3s.

For now, change the callers to pass in `context.TODO()` as a
placeholder.