example.go: use a context that can be canceled by signals
This context currently isn't actually used after initialization, but
the most plausible way for it to be used in the future is to terminate
the app on signals. So set that up.

Technically, these signals are not are not guaranteed to be present on
all platforms, and I assume that one or both of them are missing on
Windows. OTOH, AFAIK Home Assistant only runs on Linux variants, and
after all this is just an example program, so I think that it's OK.