Ensure that message IDs are used in the correct order
The websocket API insists that message IDs be monotonically
increasing (i.e., not just unique). That means that we need to hold a
lock the whole time from when we allocate the message ID until we send
the message. Change the interface for sending messages to make this
safe:

* Move the "write" methods of `Conn` to a new type, `LockedConn`.

* Add a method `Conn.Send()`, which takes a callback function as an
  argument. The callback gets a `LockedConn` as an argument. (This is
  the only way to get a hold of a `LockedConn`.) The callback is
  invoked while `Conn.writeLock` is held.

The callback can thus allocate new message IDs and send messages with
those IDs without fear that another thread interferes with it.