LockedConn: add a mechanism for subscribing to particular message IDs
It will be easier to route messages to the correct subscriber, and
prevent subscribers from interfering with each other, if subscribing
at the `Conn` level is done by message ID. Add a mechanism for doing
so.

For now, messages are also passed to a `defaultSubscriber`. That will
soon go away.