App.Start(): move more of the message handling into the subscriber
The old code wrote incoming messages to a channel, and then (in a
separate goroutine) read and processed the messages from the channel.
Note that the second goroutine created yet another goroutine for
processing each message.

Instead, write a `Subscriber` that processes the messages directly
(while continuing to create a goroutine for each message). Pass this
subscriber to `Conn.ListenWebsocket()` as the default subscriber.

In future commits, we will add more specific subscribers for each type
of message.