ensure write operations are done on a pointer
DailySchedule.updateNextRunTime took a copy and tried to write to it, which caused the loop to run infinitely since the next run time was never updated. This led to runaway memory usage and crashed my linux machine.

Interval did not have the problem even though it is also writing to a copy, because it does so in the same method where it reads. I moved that write to a separate method to prevent this same bug from being accidentally introduced in the future.