CallServiceMessage: new type
Introduce a new type, `CallServiceMessage`, to contain the entire
message required to invoke an HA service. This type embeds an instance
of `BaseServiceRequest`.

Remove the `ID` and `Type` fields from `BaseServiceRequest`, because
they don't need to be set by the caller, but are rather managed within
`App.Call()`. Move these fields to `BaseServiceRequest` by embedding a
`websocket.BaseMessage` instance the the latter type.