App.Call(): new method
Add a new version of method `App.Call()` that not only invokes a Home
Assistant API (like `CallAndForget()`), but also waits for the server
to respond and unmarshals the result into a caller-provided variable.
If the server returns an error, return that error as a
`*websocket.ResultError`.

This makes it easy to call APIs that return results, and also makes it
easy for the caller to detect when the requested action failed.