services: return the result of the service call
When calling a service, use `Call()` rather than `CallAndForget()` so
that the result of the service call is collected. Return the result to
the caller as type `any`. (If people go to the trouble of figuring out
the structure of the methods' responses, these return types can be
made more specific.

This requires a `context.Context` argument to be added to the service
call arguments. It can be used, for example, to limit the amount of
time to wait for a response from the server.